+++ NEWS.md
@@ -3,6 +3,8 @@
 * An error is produced if a too large autorregresive lag is used.
 * An error is produced in method KNN when k is greater than the size of the
   training set.
+* A warning is produced when the time series is too short to estimate
+  forecast accuracy.
 
 # utsf 1.1.0